+++ typescript/infra/config/environments/mainnet3/agent.ts
@@ -7,7 +7,10 @@ import {
   ChainName,
   GasPaymentEnforcement,
   GasPaymentEnforcementPolicyType,
+  IsmCacheConfig,
+  IsmCachePolicy,
   MatchingList,
+  ModuleType,
   RpcConsensusType,
 } from '@hyperlane-xyz/sdk';
 
@@ -29,7 +32,7 @@ import { ALL_KEY_ROLES, Role } from '../../../src/roles.js';
 import { Contexts } from '../../contexts.js';
 import { getDomainId } from '../../registry.js';
 
-import { environment } from './chains.js';
+import { environment, ethereumChainNames } from './chains.js';
 import { helloWorld } from './helloworld.js';
 import aaveSenderAddresses from './misc-artifacts/aave-sender-addresses.json';
 import everclearSenderAddresses from './misc-artifacts/everclear-sender-addresses.json';
@@ -71,7 +74,6 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     artela: true,
     arthera: true,
     astar: true,
-    astarzkevm: true,
     aurora: true,
     avalanche: true,
     b3: true,
@@ -91,7 +93,9 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     conwai: true,
     coredao: true,
     corn: true,
+    coti: true,
     cyber: true,
+    deepbrainchain: true,
     degenchain: true,
     dogechain: true,
     duckchain: true,
@@ -117,7 +121,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     hyperevm: true,
     immutablezkevmmainnet: true,
     inevm: true,
-    infinityvm: true,
+    infinityvm: false,
     injective: true,
     ink: true,
     kaia: true,
@@ -140,7 +144,9 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     morph: true,
     nero: true,
     neutron: true,
+    nibiru: true,
     oortmainnet: true,
+    opbnb: true,
     optimism: true,
     orderly: true,
     osmosis: true,
@@ -151,6 +157,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     prom: true,
     proofofplay: true,
     rarichain: true,
+    reactive: true,
     real: true,
     redstone: true,
     rivalz: true,
@@ -207,7 +214,6 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     artela: true,
     arthera: true,
     astar: true,
-    astarzkevm: true,
     aurora: true,
     avalanche: true,
     b3: true,
@@ -227,7 +233,9 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     conwai: true,
     coredao: true,
     corn: true,
+    coti: true,
     cyber: true,
+    deepbrainchain: true,
     degenchain: true,
     dogechain: true,
     duckchain: true,
@@ -253,7 +261,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     hyperevm: true,
     immutablezkevmmainnet: true,
     inevm: true,
-    infinityvm: true,
+    infinityvm: false,
     injective: true,
     ink: true,
     kaia: true,
@@ -276,7 +284,9 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     morph: true,
     nero: true,
     neutron: true,
+    nibiru: true,
     oortmainnet: true,
+    opbnb: true,
     optimism: true,
     orderly: true,
     osmosis: true,
@@ -287,6 +297,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     prom: true,
     proofofplay: true,
     rarichain: true,
+    reactive: true,
     real: true,
     redstone: true,
     rivalz: true,
@@ -343,7 +354,6 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     artela: true,
     arthera: true,
     astar: true,
-    astarzkevm: true,
     aurora: true,
     avalanche: true,
     b3: true,
@@ -363,7 +373,9 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     conwai: true,
     coredao: true,
     corn: true,
+    coti: true,
     cyber: true,
+    deepbrainchain: true,
     degenchain: true,
     dogechain: true,
     duckchain: true,
@@ -389,7 +401,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     hyperevm: true,
     immutablezkevmmainnet: true,
     inevm: true,
-    infinityvm: true,
+    infinityvm: false,
     ink: true,
     injective: true,
     kaia: true,
@@ -412,7 +424,9 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     morph: true,
     nero: true,
     neutron: true,
+    nibiru: true,
     oortmainnet: true,
+    opbnb: true,
     optimism: true,
     orderly: true,
     osmosis: true,
@@ -423,6 +437,7 @@ export const hyperlaneContextAgentChainConfig: AgentChainConfig<
     prom: true,
     proofofplay: true,
     rarichain: true,
+    reactive: true,
     real: true,
     redstone: true,
     rivalz: true,
@@ -720,15 +735,24 @@ const blacklistedMessageIds = [
 
 // Blacklist matching list intended to be used by all contexts.
 const blacklist: MatchingList = [
-  {
-    // Eco, who's sending a lot of messages not intended to be processed by the relayer.
-    // A temporary measure to prevent some wasted effort on our relayer.
-    senderAddress: '0xd890d66a0e2530335D10b3dEb5C8Ec8eA1DaB954',
-  },
   ...blacklistedMessageIds.map((messageId) => ({
     messageId,
   })),
 ];
+
+const defaultIsmCacheConfig: IsmCacheConfig = {
+  // Default ISM Routing ISMs change configs based off message content,
+  // so they are not specified here.
+  moduleTypes: [
+    ModuleType.AGGREGATION,
+    ModuleType.MERKLE_ROOT_MULTISIG,
+    ModuleType.MESSAGE_ID_MULTISIG,
+  ],
+  // SVM is explicitly not cached as the default ISM is a multisig ISM
+  // that routes internally.
+  chains: ethereumChainNames,
+  cachePolicy: IsmCachePolicy.IsmSpecific,
+};
 
 const hyperlane: RootAgentConfig = {
   ...contextBase,
@@ -739,17 +763,19 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: 'af73f39-20250325-130319',
+      tag: 'ef039ae-20250411-104801',
     },
     blacklist,
     gasPaymentEnforcement: gasPaymentEnforcement,
     metricAppContextsGetter,
+    defaultIsmCacheConfig,
+    allowContractCallCaching: true,
     resources: relayerResources,
   },
   validators: {
     docker: {
       repo,
-      tag: 'f5174e6-20250310-182921',
+      tag: '8ab7c80-20250326-191115',
     },
     rpcConsensusType: RpcConsensusType.Quorum,
     chains: validatorChainConfig(Contexts.Hyperlane),
@@ -760,7 +786,7 @@ const hyperlane: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: 'a76dd09-20250325-111203',
+      tag: '8ab7c80-20250326-191115',
     },
     resources: scraperResources,
   },
@@ -775,7 +801,7 @@ const releaseCandidate: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: 'af73f39-20250325-130319',
+      tag: 'ef039ae-20250411-104801',
     },
     blacklist,
     // We're temporarily (ab)using the RC relayer as a way to increase
@@ -783,6 +809,8 @@ const releaseCandidate: RootAgentConfig = {
     // whitelist: releaseCandidateHelloworldMatchingList,
     gasPaymentEnforcement,
     metricAppContextsGetter,
+    defaultIsmCacheConfig,
+    allowContractCallCaching: true,
     resources: relayerResources,
   },
   validators: {
@@ -809,11 +837,13 @@ const neutron: RootAgentConfig = {
     rpcConsensusType: RpcConsensusType.Fallback,
     docker: {
       repo,
-      tag: 'af73f39-20250325-130319',
+      tag: 'ef039ae-20250411-104801',
     },
     blacklist,
     gasPaymentEnforcement,
     metricAppContextsGetter,
+    defaultIsmCacheConfig,
+    allowContractCallCaching: true,
     resources: relayerResources,
   },
 };